Splitted the main script into root and user parts, added some checks
All the checks are done while in user mode, in this way if something
fails we don't even bother the user asking him the root/sudo
password. If everything is fine then the root password is needed only
to actually bring up the TUN/TAP interface. Doing so, the removal of a
connection is entirely done in user mode.