Implemented signals in ConfigurationPage
Major refactoring, the ConfigurationPage widget now sends signals
related to the connection status, in this way we're no longer forced
to link every ConfigurationPage to its father (thus creating lot of
useless pointers that may confuse the garbage collector) and it's
overall a cleaner solution. The notification bubbles class handling is
now stuff for the main window (as it should have always been) and the
VDEConnector is a Singleton (as it should have always been).